+++ dto/global_factor.go
@@ -1,4 +1,4 @@
-package constant
+package dto
 
 import "time"
